fixed issues with WatchListButton, added/updated the TODO's for Merve
